+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.1.0
 
 * `[New]` Implement RADIUS authentication.
+* `[New]` Implement Regular Expression matching for zones.
 * `[Fix]` Update password complexity checks to accept non-7-bit ASCII.