+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 
 * `[New]` Implement RADIUS authentication.
 * `[New]` Implement Regular Expression matching for zones.
+* `[New]` Optional functionality to change user password on Active Directory via LDAP.
 * `[Fix]` Update password complexity checks to accept non-7-bit ASCII.
+* `[Fix]` Fix 2FA secrets 160-bit key issue.